backend/requirements.txt
fastapi
uvicorn
pydantic
passlib
python-jose[cryptography]
PyJWT
